// apps/portal-api/src/model/kanel/public/RolePortal.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.RolePortal */
export type RolePortalId = string & { __brand: 'RolePortalId' };

/** Represents the table public.RolePortal */
export default interface RolePortal {
  id: RolePortalId;

  name: string | null;
}

/** Represents the initializer for the table public.RolePortal */
export interface RolePortalInitializer {
  id: RolePortalId;

  name?: string | null;
}

/** Represents the mutator for the table public.RolePortal */
export interface RolePortalMutator {
  id?: RolePortalId;

  name?: string | null;
}